requirements.txt
amqp==5.2.0
billiard==4.2.0
blinker==1.8.2
cachelib==0.9.0
celery==5.4.0
cfgv==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distlib==0.3.8
filelock==3.14.0
Flask==3.0.3
Flask-Caching==2.3.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flower==2.0.1
greenlet==3.0.3
gunicorn==22.0.0
humanize==4.9.0
identify==2.5.36
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.3.7
MarkupSafe==2.1.5
nodeenv==1.9.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pip-autoremove==0.10.0
platformdirs==4.2.2
pre-commit==3.7.1
prometheus_client==0.20.0
prompt_toolkit==3.0.45
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
SQLAlchemy==2.0.30
tornado==6.4
typing_extensions==4.12.0
tzdata==2024.1
Unidecode==1.3.8
vine==5.1.0
virtualenv==20.26.2
wcwidth==0.2.13
Werkzeug==3.0.3
WTForms==3.1.2
XlsxWriter==3.2.0
